Riskgroup is now the last page of registration
